refactor: 타입 파일 추가 및 기존 파일에서의 타입 경로 수정
