+++ src/core/writer/local-doc-writer.ts
@@ -1,7 +1,8 @@
 import * as fs from "node:fs/promises";
 import path from "node:path";
 
-import { DocWriterConfig } from "./doc-writer-config";
+import { DocWriterConfig } from "../../types/doc-writer-config.type";
+
 import { DocWriter } from "./doc-writer.interface";
 
 export class LocalDocWriter implements DocWriter {
