+++ src/types/descriptors.type.ts
