+++ src/types/doc-options.type.ts
@@ -1,4 +1,4 @@
-import { FieldDescriptor } from "./descriptors";
+import { FieldDescriptor } from "./descriptors.type";
 
 /** 문서 포맷 타입 */
 export type DocumentFormat = "adoc" | "md";
