+++ src/types/doc-writer-config.type.ts
