+++ src/types/http-trace.type.ts
@@ -1,6 +1,6 @@
 import { Response as SupertestResponseType } from "supertest";
 
-import { HttpMethod } from "./http";
+import { HttpMethod } from "./http.type";
 
 export type SupertestResponse = SupertestResponseType;
