+++ src/types/snippets.type.ts
